+++ functions/_cities/presov.ts
@@ -14,8 +14,7 @@ const config: CityConfig = {
     bounds: [21.13, 48.93, 21.37, 49.08],
     feed: {
         realtimeUrl: 'https://egov.presov.sk/geodatakatalog/dpmp.csv',
-        staticDataUrl: 'https://data.departs.app',
-        vehicleMetadataFile: 'dpmp-vehicles.json?v=2'
+        staticDataUrl: 'https://data.departs.app'
     },
     isBeta: true,
     filters: {